# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type RideStatus = (typeof RideStatus)[keyof typeof RideStatus];

export const RideStatus = {
  active: "active",
  accepted: "accepted",
} as const;

export interface CreateRideRequest {
  /** @minLength 2 */
  origin: string;
  /** @minLength 2 */
  destination: string;
  /** @minLength 3 */
  pickupAddress: string;
  /** @maxLength 500 */
  notes?: string;
  /**
   * @minimum 1
   * @maximum 7
   */
  seats: number;
  departAfter: string;
  departBefore: string;
}

export interface AcceptRideRequest {
  /** @minLength 2 */
  driverName: string;
  /** @minLength 5 */
  driverPhone: string;
}

export interface RideRequest {
  id: string;
  origin: string;
  destination: string;
  pickupAddress: string;
  notes: string | null;
  seats: number;
  route: string;
  status: RideStatus;
  driverName: string | null;
  driverPhone: string | null;
  departAfter: string;
  departBefore: string;
  createdAt: string;
  acceptedAt: string | null;
}

export interface RideStats {
  activeRequests: number;
  acceptedRequests: number;
  totalSeats: number;
}