# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /rides-api
    description: Rides API path
tags:
  - name: health
    description: Health operations
  - name: rides
    description: Ride request operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /requests:
    get:
      operationId: listRideRequests
      tags: [rides]
      summary: List active passenger requests
      responses:
        "200":
          description: Active requests
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/RideRequest"
    post:
      operationId: createRideRequest
      tags: [rides]
      summary: Create a passenger ride request
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateRideRequest"
      responses:
        "201":
          description: Created request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RideRequest"
  /requests/{id}:
    get:
      operationId: getRideRequest
      tags: [rides]
      summary: Get one ride request
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Ride request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RideRequest"
        "404":
          description: Request not found
  /requests/{id}/accept:
    post:
      operationId: acceptRideRequest
      tags: [rides]
      summary: Driver accepts a passenger request
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AcceptRideRequest"
      responses:
        "200":
          description: Accepted request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RideRequest"
        "404":
          description: Request not found
        "409":
          description: Request already accepted
  /stats:
    get:
      operationId: getRideStats
      tags: [rides]
      summary: Small dashboard summary
      responses:
        "200":
          description: Ride statistics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RideStats"
components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    RideStatus:
      type: string
      enum: [active, accepted]
    CreateRideRequest:
      type: object
      properties:
        origin:
          type: string
          minLength: 2
        destination:
          type: string
          minLength: 2
        pickupAddress:
          type: string
          minLength: 3
        notes:
          type: string
          maxLength: 500
        seats:
          type: integer
          minimum: 1
          maximum: 7
        departAfter:
          type: string
          format: date-time
        departBefore:
          type: string
          format: date-time
      required:
        - origin
        - destination
        - pickupAddress
        - seats
        - departAfter
        - departBefore
    AcceptRideRequest:
      type: object
      properties:
        driverName:
          type: string
          minLength: 2
        driverPhone:
          type: string
          minLength: 5
      required:
        - driverName
        - driverPhone
    RideRequest:
      type: object
      properties:
        id:
          type: string
        origin:
          type: string
        destination:
          type: string
        pickupAddress:
          type: string
        notes:
          type: string
          nullable: true
        seats:
          type: integer
        route:
          type: string
        status:
          $ref: "#/components/schemas/RideStatus"
        driverName:
          type: string
          nullable: true
        driverPhone:
          type: string
          nullable: true
        departAfter:
          type: string
          format: date-time
        departBefore:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
        acceptedAt:
          type: string
          format: date-time
          nullable: true
      required:
        - id
        - origin
        - destination
        - pickupAddress
        - notes
        - seats
        - route
        - status
        - driverName
        - driverPhone
        - departAfter
        - departBefore
        - createdAt
        - acceptedAt
    RideStats:
      type: object
      properties:
        activeRequests:
          type: integer
        acceptedRequests:
          type: integer
        totalSeats:
          type: integer
      required:
        - activeRequests
        - acceptedRequests
        - totalSeats