# lib/api-zod/src/generated/types/createRideRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CreateRideRequest {
  /** @minLength 2 */
  origin: string;
  /** @minLength 2 */
  destination: string;
  /** @minLength 3 */
  pickupAddress: string;
  /** @maxLength 500 */
  notes?: string;
  /**
   * @minimum 1
   * @maximum 7
   */
  seats: number;
  departAfter: Date;
  departBefore: Date;
}