# lib/api-zod/src/generated/types/rideRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { RideStatus } from "./rideStatus";

export interface RideRequest {
  id: string;
  origin: string;
  destination: string;
  pickupAddress: string;
  notes: string | null;
  seats: number;
  route: string;
  status: RideStatus;
  driverName: string | null;
  driverPhone: string | null;
  departAfter: Date;
  departBefore: Date;
  createdAt: Date;
  acceptedAt: Date | null;
}